Soon release

+++ src/warlock/game/Game.java
@@ -7,24 +7,33 @@ package warlock.game;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.Iterator;
-import org.lwjgl.input.Keyboard;
 import warlock.camera.Camera;
+import warlock.constant.ZLayers;
+import warlock.graphic.Color;
 import warlock.graphic.Graphic;
 import warlock.input.InputHandler;
+import warlock.lobby.Lobby;
+import warlock.object.character.DeathListener;
+import warlock.player.ObserverPlayer;
 import warlock.player.Player;
 
 /**
- * -----> TODO: Add listener from player to listen when kills etc happen to display messages
  *
  * @author Focusrite
  */
-public class Game {
+public class Game implements DeathListener {
 
+   private static final double TRANSITION_TIME = 2;
+   private double transition = 0;
+   private boolean inTransition = true;
+   private static final int TRANSITION_BARS = 16;
    private ArrayList<Player> playerList = new ArrayList<>();
-   private ArrayList<Player> scoreTable = new ArrayList<>();
+   private ArrayList<Player> scoretable = new ArrayList<>();
    private Player player; //The local player
    private GamePhase phase;
+   private GamePhase temporaryPhase;
    private Camera camera;
+   private int firstTo;
 
    public Player getPlayer() {
       return player;
@@ -42,12 +51,45 @@ public class Game {
       return playerList.listIterator();
    }
 
-   public Game(Camera camera) {
+   public ArrayList<Player> getScoreTable() {
+      return scoretable;
+   }
+
+   public Game(Camera camera, Lobby lobby) {
       this.camera = camera;
+      readFromLobby(lobby);
+   }
+
+   private void readFromLobby(Lobby lobby) {
+      if (lobby.getSelfPlayer() == null) {
+         setPlayer(new ObserverPlayer());
+      }
+      else {
+         setPlayer(lobby.getSelfPlayer());
+      }
+      for (int i = 0; i < lobby.getLobbyList().length; i++) {
+         if (lobby.getLobbyList()[i] != null) {
+            addPlayer(lobby.getLobbyList()[i]);
+         }
+      }
+      setFirstTo(lobby.getFirstTo());
+   }
+
+   public int getFirstTo() {
+      return firstTo;
+   }
+
+   public void setFirstTo(int firstTo) {
+      this.firstTo = firstTo;
    }
 
    public void init() {
-      setPhase(new PlayPhase(this)); //should start at shopphase
+      if(isObservermode()) {
+         setPhase(new PlayPhase(this));
+      }
+      else {
+         setPhase(new ShopPhase(this));
+      }
    }
 
    public GamePhase getPhase() {
@@ -55,11 +97,29 @@ public class Game {
    }
 
    public final void setPhase(GamePhase phase) {
+      temporaryPhase = phase;
+      if (this.phase == null) {
+         changedPhase();
+      }
+      else {
+         transition = TRANSITION_TIME;
+         inTransition = true;
+      }
+   }
+
+   private boolean changedPhase() {
+      if (temporaryPhase == null || transition > 0) {
+         return false;
+      }
       if (this.phase != null) {
+         transition = TRANSITION_TIME;
+         inTransition = !inTransition;
          this.phase.terminate();
       }
-      this.phase = phase;
+      this.phase = temporaryPhase;
       this.phase.init();
+      this.temporaryPhase = null;
+      return true;
    }
 
    public void addPlayer(Player p) {
@@ -67,25 +127,64 @@ public class Game {
          this.playerList.remove(p); //Same playerId exist, remove to replace with new player
       }
       this.playerList.add(p);
-      this.scoreTable.add(p);
-      Collections.sort(this.scoreTable);
+      this.scoretable.add(p);
+      p.getWarlock().addDeathListener(this); //Register KO listener
+      Collections.sort(this.scoretable);
    }
 
    public void modifyScore(Player p, int dScore) {
       p.modifyScore(dScore);
-      Collections.sort(this.scoreTable);
+      Collections.sort(this.scoretable);
    }
 
    public void update(double dt) {
+      if (changedPhase()) {
+         return;
+      }
+      if (transition > 0) {
+         transition -= dt;
+      }
       phase.update(dt);
    }
 
    public void render(Graphic g) {
+      if (transition > 0) {
+         drawTransition(g);
+      }
       phase.render(g);
    }
 
+   private void drawTransition(Graphic g) {
+      g.setScreenCoordinates(true);
+      double transitionLeft = (inTransition) ? 1 - (transition / TRANSITION_TIME) : (transition / TRANSITION_TIME);
+      int x = (g.getScreenWidth() / TRANSITION_BARS) / 2;
+      double height = (transitionLeft * TRANSITION_BARS);
+      while (height > 0) {
+         int h = (int) (g.getScreenHeight() * Math.min(1, height));
+         g.drawRectangle(x, h / 2, ZLayers.OVERLAY,
+            g.getScreenWidth() / TRANSITION_BARS, h,
+            0, Color.BLACK);
+         x += (g.getScreenWidth() / TRANSITION_BARS);
+         height -= 1;
+      }
+      g.setScreenCoordinates(false);
+   }
+
    public void handleInput(InputHandler input) {
-      this.player.handleInput(input);
       this.phase.handleInput(input);
+   }
+
+   @Override
+   public void notifyDeath(Player p, Player killer) {
+      this.phase.notifyDeath(p, killer);
+      Collections.sort(this.scoretable);
+      if(scoretable.get(0).getScore() >= firstTo && scoretable.size() > 1 &&
+         scoretable.get(0).getScore() != scoretable.get(1).getScore()) {
+         setPhase(new GameOverPhase(this, scoretable));
+      }
+   }
+
+   boolean isObservermode() {
+      return getPlayer() instanceof ObserverPlayer;
    }
 }
